requirements.txt
absl-py==1.4.0
accelerate==0.18.0
ai2thor==2.1.0
antlr4-python3-runtime==4.9.3
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asttokens==2.2.1
attrs==21.4.0
backcall==0.2.0
backports.functools-lru-cache==1.6.4
beautifulsoup4==4.11.0
bitsandbytes==0.37.2
black==22.3.0
bleach==5.0.0
blis==0.7.9
brotlipy==0.7.0
cachetools==5.3.0
catalogue==2.0.8
certifi==2022.12.7
cffi==1.14.6
charset-normalizer==2.0.12
click==8.1.2
cloudpickle==3.0.0
comm==0.1.3
confection==0.0.4
cryptography==39.0.0
cycler==0.11.0
cymem==2.0.7
cytoolz==0.12.1
debugpy==1.6.6
decorator==4.4.2
defusedxml==0.7.1
dill==0.3.6
docker-pycreds==0.4.0
entrypoints==0.4
executing==1.2.0
fastjsonschema==2.15.3
filelock==3.6.0
Flask==2.1.1
fonttools==4.31.2
future==0.18.2
gdown==4.7.1
gitdb==4.0.9
GitPython==3.1.27
gym==0.26.2
gym-notices==0.0.8
h5py==3.6.0
h5py-wrapper==1.1.0
hdfdict==0.3.1
huggingface-hub==0.13.3
hydra-core==1.3.2
idna==3.4
imageio==2.19.3
imageio-ffmpeg==0.4.7
importlib-metadata==4.11.3
ipykernel==6.22.0
ipython==8.11.0
ipython-genutils==0.2.0
ipywidgets==7.7.0
itsdangerous==2.1.2
jedi==0.18.2
jellyfish==0.10.0
Jinja2==3.1.1
joblib==1.1.0
jsonschema==4.4.0
jupyter==1.0.0
jupyter_client==8.1.0
jupyter-console==6.4.3
jupyter_core==5.3.0
jupyterlab-pygments==0.1.2
jupyterlab-widgets==1.1.0
kiwisolver==1.4.2
langcodes==3.3.0
llvmlite==0.38.0
lmdb==1.3.0
MarkupSafe==2.1.1
matplotlib==3.5.1
matplotlib-inline==0.1.6
mistune==0.8.4
mkl-fft==1.3.6
mkl-random==1.2.2
mkl-service==2.4.0
ml-pyxis==0.4.dev0
moviepy==1.0.3
msgpack==1.0.3
multiprocess==0.70.14
murmurhash==1.0.9
mypy-extensions==0.4.3
nbclient==0.5.13
nbconvert==6.4.5
nbformat==5.3.0
nest-asyncio==1.5.6
networkx==2.7.1
nltk==3.7
notebook==6.4.10
numpy==1.22.3
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
omegaconf==2.3.0
opencv-python==4.5.5.64
packaging==21.3
pandas==1.4.1
pandocfilters==1.5.0
parso==0.8.3
pathos==0.3.0
pathspec==0.9.0
pathtools==0.1.2
pathy==0.10.1
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.1.0
pip==23.0.1
platformdirs==2.5.1
plotly==5.9.0
pox==0.3.2
ppft==1.7.6.6
preshed==3.0.8
proglog==0.1.10
progressbar2==4.0.0
prometheus-client==0.14.0
promise==2.3
prompt-toolkit==3.0.38
protobuf==3.20.0
psutil==5.9.4
ptyprocess==0.7.0
pure-eval==0.2.2
pycparser==2.21
pydantic==1.10.7
Pygments==2.14.0
pyOpenSSL==23.1.0
pyparsing==3.0.7
pyphen==0.14.0
pyrsistent==0.18.1
PySocks==1.7.1
pytest-runner==6.0.0
python-dateutil==2.8.2
python-utils==3.1.0
pytz==2022.1
pyxis==0.2.1
PyYAML==5.4.1
pyzmq==25.0.2
qtconsole==5.3.0
QtPy==2.0.1
regex==2022.3.15
requests==2.28.2
revtok==0.0.3
sacremoses==0.0.49
scikit-learn==1.0.2
scipy==1.8.0
seaborn==0.11.2
Send2Trash==1.8.0
sentence-transformers==2.2.2
sentencepiece==0.1.96
sentry-sdk==1.5.8
setproctitle==1.2.2
setuptools==62.0.0
shortuuid==1.0.8
six==1.16.0
smart-open==6.3.0
smmap==5.0.0
soupsieve==2.3.2
srsly==2.4.6
stack-data==0.6.2
tenacity==8.0.1
tensorboardX==1.8
terminado==0.13.3
testpath==0.6.0
textacy==0.13.0
thinc==8.1.9
threadpoolctl==3.1.0
tokenizers==0.13.3
tomli==2.0.1
toolz==0.12.0
torch==1.13.1
torchvision==0.14.1
tornado==6.2
tqdm==4.63.1
traitlets==5.9.0
transformers==4.28.1
typer==0.7.0
typing_extensions==4.5.0
urllib3==1.26.9
vocab==0.0.5
wandb==0.12.14
wasabi==1.1.1
wcwidth==0.2.6
webencodings==0.5.1
Werkzeug==2.1.1
wheel==0.40.0
widgetsnbextension==3.6.0
zipp==3.7.0
